Refactor product management components: remove ProductsList and PromoteProductForm, add CreateProductForm, EditProduct, ProductDetails, and updated ProductsList with pagination.
